Modify packages
Kepp connectionPooler in cluster package itself. This helps in getting
rid of previously added packages and interfaces.
Based on the new structure of ConnectionPooler in cluster, there are
functional changes to come.
This commit is only for updating and cleaning of packages.
